refactored the OM to use the event sourcing pattern
I refer to the event sourcing pattern as described here:
https://msdn.microsoft.com/en-us/library/dn589792.aspx

This change is a big one.  There were three driving reasons behind this
change:

1) Identifying changes that a user hasn't seen when returning to a
review and examining the history of the review are both easy to
accomplish.
2) Network synchronization is easy if change events are the atomic unit
of record used internally. Simply compare event IDs to see what changes
are missing on either end of a network pipe.
3) Simple persistence if an entire review can be summarized as an array
of of plain-old-data JSON objects. This is easy to
read/write/parse/compress and view in a text editor without external
libraries or tooling required.

Also as part of this change I added a simple utility method to generate
high-resolution timestamps for use in uniquely identifying change events
in event logs. The returned values are accurate down to approximately
1/1000th of a millisecond using the performance.now() timer. No further
precision can be represented in JavaScript's double, but that seems
sufficient for the needs of the log.